Add an option to the root config to include a projectConfigFilePath
This option which takes a glob to represent the project config file path and a name property to identify a project name. If there is no name field, we guess the project's name based on the parent directory of the config file. I pulled in es-promisify to make this change since we still running CI against node v6 and I updated glob.